registry classes unification
